+++ CHANGELOG.md
@@ -1,16 +1,29 @@
 # Change Log
+
 Only partial changelog, [commit history](https://framagit.org/SondagePro-LimeSurvey-plugin/sendMailCron/commits/master) show all changelog.
 
+## Unreleased
+
+### Fix
+
+- Add beforeTokenEmail event when send an email
+
 ## [0.2.0] - 2016-12-14
+
 - This version is ditributed in Affero General Public License
-### Feature ###
+
+### Featur
+
 - Allow different settings by survey
 - Add a global batch file size
 
 ## [0.1.1] - 2016-09-17
-### Fix ###
+
+### Fix
+
 - trim email before validate
 
 ## [0.0.1] - 2016-05-17
-### Feature ###
+
+### Feature
 - send email by cron event